+++ test/fixtures/postgres/addresses.sql
@@ -1,5 +1,5 @@
-create table addresses (
-  id uuid not null constraint addresses_pk primary key,
-  customer_id int constraint addresses_customers_id_fk references customers,
-  city varchar not null
+CREATE TABLE addresses (
+  id VARCHAR(36) NOT NULL CONSTRAINT addresses_pk PRIMARY KEY,
+  customer_id INT CONSTRAINT addresses_customers_id_fk REFERENCES customers,
+  city VARCHAR(30) NOT NULL UNIQUE
 );